plugin-assistant: move ConnectorAuthMenu component out of plugin-connector
The standalone connector-auth menu lives with its only consumer (IntegrationPrompt)
in plugin-assistant; plugin-connector keeps and exports `connectorAuthActions` and
`CONNECTOR_AUTH_GROUP_ID` from its main entry.
