+++ packages/plugins/plugin-assistant/src/components/ConnectorAuthMenu/ConnectorAuthMenu.tsx
@@ -7,15 +7,13 @@ import React, { useContext, useMemo } from 'react';
 
 import { useCapabilities } from '@dxos/app-framework/ui';
 import { type Database, Filter, type Obj, type Ref } from '@dxos/echo';
+import { Connection, Connector, CONNECTOR_AUTH_GROUP_ID, connectorAuthActions } from '@dxos/plugin-connector';
 import { Graph, useActionRunner } from '@dxos/plugin-graph';
 import { useQuery } from '@dxos/react-client/echo';
 import { IconButton, useTranslation } from '@dxos/react-ui';
 import { Menu, useGraphMenuActions } from '@dxos/react-ui-menu';
 
 import { meta } from '#meta';
-import { Connection, Connector } from '#types';
-
-import { CONNECTOR_AUTH_GROUP_ID, connectorAuthActions } from '../../util';
 
 /** Root node the connector-auth group hangs off in the component's local graph. */
 const NODE_ID = 'connector-auth-root';
