+++ packages/plugins/plugin-assistant/src/components/ConnectorAuthMenu/index.ts
