+++ packages/plugins/plugin-connector/src/components/index.ts
@@ -2,5 +2,4 @@
 // Copyright 2025 DXOS.org
 //
 
-export * from './ConnectorAuthMenu';
 export * from './ConnectionView';
