+++ packages/plugins/plugin-connector/src/index.ts
@@ -3,7 +3,7 @@
 //
 
 export { ATMOSPHERE_PROVIDER_ID, ATMOSPHERE_SOURCE, ATPROTO_OAUTH_SCOPES } from './constants';
-export { type ConnectorAuthActionsOptions, connectorAuthActions } from './util';
+export { CONNECTOR_AUTH_GROUP_ID, type ConnectorAuthActionsOptions, connectorAuthActions } from './util';
 export { ConnectionTestError } from './errors';
 export * from './meta';
 export * from './types';
